feat(detect): Amelioration du code
